Fix: Data model alignment for SplitTimeEntity (Float to Double)

+++ app/src/main/java/com/ex/mountaintimer/SplitTimeEntity.kt
@@ -17,6 +17,6 @@ data class SplitTimeEntity(
     val checkpointIndex: Int,     // 對應 Checkpoint.index
     val checkpointName: String,   // 對應 Checkpoint.name
     val timeMs: Long,             // 從起點開始計算的時間（ms）
-    val speed: Float = 0f,        // 通過時的速度 (m/s)
-    val gForce: Float = 0f        // 通過時的總 G 力
+    val speed: Double = 0.0,      // 通過時的速度 (m/s)
+    val gForce: Double = 0.0      // 通過時的總 G 力
 )
